modularized API resources and schemas
