+++ minance/user/schemas.py
@@ -1,0 +1,20 @@
+from minance import ma
+from minance.models import User
+
+class UserSchema(ma.Schema):
+  class Meta:
+    fields = (
+      "username",
+      "email",
+      "_links"
+    )
+
+    ordered = True
+    model = User
+
+  _links = ma.Hyperlinks(
+    {
+      "self": ma.URLFor("api.user_detail", values=dict(user_id="<id>")),
+      "collection": ma.URLFor("api.users"),
+    }
+  )
